Move staging funcs to separate file
